+++ service/history/api/respondworkflowtaskcompleted/workflow_task_completed_handler.go
@@ -55,6 +55,7 @@ import (
 	"go.temporal.io/server/common/tasktoken"
 	"go.temporal.io/server/internal/effect"
 	"go.temporal.io/server/internal/protocol"
+	"go.temporal.io/server/service/history/api"
 	"go.temporal.io/server/service/history/configs"
 	"go.temporal.io/server/service/history/shard"
 	"go.temporal.io/server/service/history/workflow"
@@ -85,7 +86,7 @@ type (
 		updateRegistry                  update.Registry
 
 		// validation
-		attrValidator                  *commandAttrValidator
+		attrValidator                  *api.CommandAttrValidator
 		sizeLimitChecker               *workflowSizeChecker
 		searchAttributesMapperProvider searchattribute.MapperProvider
 
@@ -124,7 +125,7 @@ func newWorkflowTaskCompletedHandler(
 	mutableState workflow.MutableState,
 	updateRegistry update.Registry,
 	effects effect.Controller,
-	attrValidator *commandAttrValidator,
+	attrValidator *api.CommandAttrValidator,
 	sizeLimitChecker *workflowSizeChecker,
 	logger log.Logger,
 	namespaceRegistry namespace.Registry,
@@ -173,7 +174,7 @@ func (handler *workflowTaskCompletedHandler) handleCommands(
 	commands []*commandpb.Command,
 	msgs *collection.IndexedTakeList[string, *protocolpb.Message],
 ) ([]workflowTaskResponseMutation, error) {
-	if err := handler.attrValidator.validateCommandSequence(
+	if err := handler.attrValidator.ValidateCommandSequence(
 		commands,
 	); err != nil {
 		return nil, err
@@ -419,7 +420,7 @@ func (handler *workflowTaskCompletedHandler) handleCommandProtocolMessage(
 
 	if err := handler.validateCommandAttr(
 		func() (enumspb.WorkflowTaskFailedCause, error) {
-			return handler.attrValidator.validateProtocolMessageAttributes(
+			return handler.attrValidator.ValidateProtocolMessageAttributes(
 				namespaceID,
 				attr,
 				executionInfo.WorkflowRunTimeout,
@@ -447,7 +448,7 @@ func (handler *workflowTaskCompletedHandler) handleCommandScheduleActivity(
 
 	if err := handler.validateCommandAttr(
 		func() (enumspb.WorkflowTaskFailedCause, error) {
-			return handler.attrValidator.validateActivityScheduleAttributes(
+			return handler.attrValidator.ValidateActivityScheduleAttributes(
 				namespaceID,
 				attr,
 				executionInfo.WorkflowRunTimeout,
@@ -610,7 +611,7 @@ func (handler *workflowTaskCompletedHandler) handleCommandRequestCancelActivity(
 ) (*historypb.HistoryEvent, error) {
 	if err := handler.validateCommandAttr(
 		func() (enumspb.WorkflowTaskFailedCause, error) {
-			return handler.attrValidator.validateActivityCancelAttributes(attr)
+			return handler.attrValidator.ValidateActivityCancelAttributes(attr)
 		},
 	); err != nil || handler.stopProcessing {
 		return nil, err
@@ -654,7 +655,7 @@ func (handler *workflowTaskCompletedHandler) handleCommandStartTimer(
 ) (*historypb.HistoryEvent, error) {
 	if err := handler.validateCommandAttr(
 		func() (enumspb.WorkflowTaskFailedCause, error) {
-			return handler.attrValidator.validateTimerScheduleAttributes(attr)
+			return handler.attrValidator.ValidateTimerScheduleAttributes(attr)
 		},
 	); err != nil || handler.stopProcessing {
 		return nil, err
@@ -677,7 +678,7 @@ func (handler *workflowTaskCompletedHandler) handleCommandCompleteWorkflow(
 
 	if err := handler.validateCommandAttr(
 		func() (enumspb.WorkflowTaskFailedCause, error) {
-			return handler.attrValidator.validateCompleteWorkflowExecutionAttributes(attr)
+			return handler.attrValidator.ValidateCompleteWorkflowExecutionAttributes(attr)
 		},
 	); err != nil || handler.stopProcessing {
 		return nil, err
@@ -732,7 +733,7 @@ func (handler *workflowTaskCompletedHandler) handleCommandFailWorkflow(
 
 	if err := handler.validateCommandAttr(
 		func() (enumspb.WorkflowTaskFailedCause, error) {
-			return handler.attrValidator.validateFailWorkflowExecutionAttributes(attr)
+			return handler.attrValidator.ValidateFailWorkflowExecutionAttributes(attr)
 		},
 	); err != nil || handler.stopProcessing {
 		return nil, err
@@ -798,7 +799,7 @@ func (handler *workflowTaskCompletedHandler) handleCommandCancelTimer(
 ) (*historypb.HistoryEvent, error) {
 	if err := handler.validateCommandAttr(
 		func() (enumspb.WorkflowTaskFailedCause, error) {
-			return handler.attrValidator.validateTimerCancelAttributes(attr)
+			return handler.attrValidator.ValidateTimerCancelAttributes(attr)
 		},
 	); err != nil || handler.stopProcessing {
 		return nil, err
@@ -828,7 +829,7 @@ func (handler *workflowTaskCompletedHandler) handleCommandCancelWorkflow(
 
 	if err := handler.validateCommandAttr(
 		func() (enumspb.WorkflowTaskFailedCause, error) {
-			return handler.attrValidator.validateCancelWorkflowExecutionAttributes(attr)
+			return handler.attrValidator.ValidateCancelWorkflowExecutionAttributes(attr)
 		},
 	); err != nil || handler.stopProcessing {
 		return nil, err
@@ -865,7 +866,7 @@ func (handler *workflowTaskCompletedHandler) handleCommandRequestCancelExternalW
 
 	if err := handler.validateCommandAttr(
 		func() (enumspb.WorkflowTaskFailedCause, error) {
-			return handler.attrValidator.validateCancelExternalWorkflowExecutionAttributes(
+			return handler.attrValidator.ValidateCancelExternalWorkflowExecutionAttributes(
 				namespaceID,
 				targetNamespaceID,
 				handler.initiatedChildExecutionsInBatch,
@@ -893,7 +894,7 @@ func (handler *workflowTaskCompletedHandler) handleCommandRecordMarker(
 ) (*historypb.HistoryEvent, error) {
 	if err := handler.validateCommandAttr(
 		func() (enumspb.WorkflowTaskFailedCause, error) {
-			return handler.attrValidator.validateRecordMarkerAttributes(attr)
+			return handler.attrValidator.ValidateRecordMarkerAttributes(attr)
 		},
 	); err != nil || handler.stopProcessing {
 		return nil, err
@@ -938,7 +939,7 @@ func (handler *workflowTaskCompletedHandler) handleCommandContinueAsNewWorkflow(
 
 	if err := handler.validateCommandAttr(
 		func() (enumspb.WorkflowTaskFailedCause, error) {
-			return handler.attrValidator.validateContinueAsNewWorkflowExecutionAttributes(
+			return handler.attrValidator.ValidateContinueAsNewWorkflowExecutionAttributes(
 				namespaceName,
 				attr,
 				handler.mutableState.GetExecutionInfo(),
@@ -1055,7 +1056,7 @@ func (handler *workflowTaskCompletedHandler) handleCommandStartChildWorkflow(
 
 	if err := handler.validateCommandAttr(
 		func() (enumspb.WorkflowTaskFailedCause, error) {
-			return handler.attrValidator.validateStartChildExecutionAttributes(
+			return handler.attrValidator.ValidateStartChildExecutionAttributes(
 				parentNamespaceID,
 				targetNamespaceID,
 				targetNamespace,
@@ -1143,7 +1144,7 @@ func (handler *workflowTaskCompletedHandler) handleCommandSignalExternalWorkflow
 
 	if err := handler.validateCommandAttr(
 		func() (enumspb.WorkflowTaskFailedCause, error) {
-			return handler.attrValidator.validateSignalExternalWorkflowExecutionAttributes(
+			return handler.attrValidator.ValidateSignalExternalWorkflowExecutionAttributes(
 				namespaceID,
 				targetNamespaceID,
 				attr,
@@ -1203,7 +1204,7 @@ func (handler *workflowTaskCompletedHandler) handleCommandUpsertWorkflowSearchAt
 	// valid search attributes for upsert
 	if err := handler.validateCommandAttr(
 		func() (enumspb.WorkflowTaskFailedCause, error) {
-			return handler.attrValidator.validateUpsertWorkflowSearchAttributes(namespace, attr)
+			return handler.attrValidator.ValidateUpsertWorkflowSearchAttributes(namespace, attr)
 		},
 	); err != nil || handler.stopProcessing {
 		return nil, err
@@ -1255,7 +1256,7 @@ func (handler *workflowTaskCompletedHandler) handleCommandModifyWorkflowProperti
 	// valid properties
 	if err := handler.validateCommandAttr(
 		func() (enumspb.WorkflowTaskFailedCause, error) {
-			return handler.attrValidator.validateModifyWorkflowProperties(namespace, attr)
+			return handler.attrValidator.ValidateModifyWorkflowProperties(namespace, attr)
 		},
 	); err != nil || handler.stopProcessing {
 		return nil, err
